Moved slack section to seperate component
